+++ xreal/code/botlib/l_script.h
@@ -165,7 +165,7 @@ typedef struct token_s
 	int             subtype;	//last read token sub type
 #ifdef NUMBERVALUE
 	unsigned long int intvalue;	//integer value
-	long double     floatvalue;	//floating point value
+	double			floatvalue;	//floating point value
 #endif							//NUMBERVALUE
 	char           *whitespace_p;	//start of white space before token
 	char           *endwhitespace_p;	//start of white space before token
@@ -235,7 +235,7 @@ void            StripSingleQuotes(char *string);
 signed long int ReadSignedInt(script_t * script);
 
 //read a possible signed floating point number
-long double     ReadSignedFloat(script_t * script);
+double			ReadSignedFloat(script_t * script);
 
 //set an array with punctuations, NULL restores default C/C++ set
 void            SetScriptPunctuations(script_t * script, punctuation_t * p);
